haval_api/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.3
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.5
django-colorfield==0.7.2
django-cors-headers==3.13.0
django-filter==22.1
django-import-export==2.8.0
django-modeltranslation==0.18.4
django-resized==1.0.2
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.0
djangorestframework-yaml==2.0.0
drf-yasg==1.20.0
et-xmlfile==1.1.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
Markdown==3.4.1
MarkupPy==1.14
MarkupSafe==2.0.1
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.9
packaging==21.0
Pillow==9.2.0
psycopg2==2.9.3
psycopg2-binary==2.9.1
pycodestyle==2.7.0
PyJWT==2.3.0
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
simplejson==3.17.3
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
uritemplate==3.0.1
urllib3==1.26.6
xlrd==2.0.1
xlwt==1.3.0
